requirements.txt
Flask==3.1.1
Werkzeug==3.1.3
Pillow==10.4.0 